move ChecksMock and MockScenarios into a single file, use :ets
